+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2025-01-22
+
+### Added
+
+- Add `defaultLanguage` setting to set the default language for the barrel file
+
+### Changed
+
+- Update the `fileExtension` setting to allow TypeScript and JavaScript file extensions
+
 ## [1.3.1] - 2025-01-19
 
 ### Fixed
@@ -61,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of the extension starter kit minimal.
 
-[unreleased]: https://github.com/ManuelGil/vscode-typescript-generator/compare/v1.3.1...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-typescript-generator/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/ManuelGil/vscode-typescript-generator/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/ManuelGil/vscode-typescript-generator/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/ManuelGil/vscode-typescript-generator/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/ManuelGil/vscode-typescript-generator/compare/v1.2.0...v1.2.1